Move Message to own module
