+++ src/message.rs
@@ -1,0 +1,18 @@
+use scalar::Scalar;
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+/// Hashed message input to an ECDSA signature.
+pub struct Message(pub Scalar);
+
+impl Message {
+    pub fn parse(p: &[u8; 32]) -> Message {
+        let mut m = Scalar::default();
+        m.set_b32(p);
+
+        Message(m)
+    }
+
+    pub fn serialize(&self) -> [u8; 32] {
+        self.0.b32()
+    }
+}
